# optimize 'sleep()' by removing keyword-arguments and closures